refactoring here and there, specially in routes module
