+++ src/estimations/routes/sequences.py
@@ -1,24 +1,13 @@
 from http import HTTPStatus
-from typing import Union
 
 from cerberus import Validator
 from flask import jsonify, make_response, request
 
 from estimations import schemas
-from organizations.exceptions import NotFound as OrganizationNotFound
+from estimations.exc import ResourceAlreadyExists
 
-from .app import estimations_app
-from .exc import ResourceAlreadyExists, SequenceNotFound, SessionNotFound
-from .models import Sequence, Session, Value
-
-
-@estimations_app.errorhandler(SequenceNotFound)
-@estimations_app.errorhandler(OrganizationNotFound)
-@estimations_app.errorhandler(SessionNotFound)
-def handle_not_found(error: Union[OrganizationNotFound, SequenceNotFound, None]):
-    return make_response(jsonify({
-        'message': str(error),
-    }), HTTPStatus.NOT_FOUND)
+from ..app import estimations_app
+from ..models import Sequence, Value
 
 
 @estimations_app.route('/sequences', methods=['GET'])
@@ -131,31 +120,3 @@ def remove_values_from_sequence(name: str):
     sequence.remove_values()
 
     return make_response(jsonify({}), HTTPStatus.NO_CONTENT)
-
-
-@estimations_app.route('/sessions/<code>', methods=['GET'])
-def get_session(code: str):
-    if not code:
-        return make_response(jsonify({
-            'message': 'Please provide the session identifier.',
-        }), HTTPStatus.NOT_FOUND)
-
-    session = Session.lookup(code)
-
-    return make_response(jsonify(session.dump()), HTTPStatus.OK)
-
-
-@estimations_app.route('/sessions/', methods=['POST'])
-def create_session():
-    payload = request.get_json()
-
-    validator = Validator()
-    if not validator.validate(payload, schemas.CREATE_SESSION):
-        return make_response(
-            jsonify(validator.errors),
-            HTTPStatus.BAD_REQUEST,
-        )
-
-    session = Session.from_data(**payload)
-
-    return make_response(jsonify(session.dump()), HTTPStatus.CREATED)
